[FE] Refactor: API 분리 (axios로직)
