+++ frontend/src/layouts/SignUpLayout/styled.ts
