+++ frontend/src/pages/AccountPage/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as S from './styled';
-import { ContainerLayout } from '@src/layout/ContainerLayout';
-import { AccountLayout } from '@src/layout/AccountLayout';
+import { ContainerLayout } from '@src/layouts/ContainerLayout';
+import { AccountLayout } from '@src/layouts/AccountLayout';
 
 const AccountPage = (): React.ReactElement => {
   return (
