+++ frontend/src/pages/SignUpPage/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as S from './styled';
-import { ContainerLayout } from '@src/layout/ContainerLayout';
-import { SignUpLayout } from '@src/layout/SignUpLayout';
+import { ContainerLayout } from '@src/layouts/ContainerLayout';
+import { SignUpLayout } from '@src/layouts/SignUpLayout';
 
 const SignUpPage = (): React.ReactElement => {
   return (
